app/requirements.txt
anyio==4.8.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
sniffio==1.3.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
waitress==3.0.2
Werkzeug==3.1.3
